ai_engine/requirements.txt
pandas
scikit-learn
mysql-connector-python
python-dotenv
